+++ crates/base_db/src/input.rs
@@ -32,7 +32,7 @@ pub struct SourceRoot {
     /// Sysroot or remote library.
     ///
     /// Libraries are considered mostly immutable, this assumption is used to
-    /// optimize salsa's query structure
+    /// optimize salsa's query structure.
     is_library: bool,
     file_set: FileSet,
 }