+++ crates/parser/src/lib.rs
@@ -1,4 +1,5 @@
-//! The parser is mostly copied from <https://github.com/arzg/eldiro/tree/master/crates/parser> with some adaptions and extensions
+//! The parser is mostly copied from <https://github.com/arzg/eldiro/tree/master/crates/parser>
+//! with some adaptions and extensions.
 
 mod cst_builder;
 mod lexer;