+++ crates/syntax/src/ast.rs
@@ -233,10 +233,10 @@ ast_node! {
     ImportSuperRelative
 }
 impl ImportSuperRelative {
-    /// Counts how often `super` appeared in a chain
+    /// Counts how often `super` appeared in a chain.
     ///
     /// # Panics
-    /// If there is a chain of more than 255 `super::`s
+    /// If there is a chain of more than 255 `super::`s.
     #[must_use]
     pub fn super_count(&self) -> u8 {
         u8::try_from(
@@ -407,7 +407,10 @@ impl EnableExtensionName {
     }
 }
 
-/// Names that can be `enable`d <https://www.w3.org/TR/WGSL/#syntax-enable_extension_name>
+#[expect(clippy::doc_paragraphs_missing_punctuation, reason = "false positive")]
+/// Names that can be `enable`d.
+///
+/// Source: <https://www.w3.org/TR/WGSL/#syntax-enable_extension_name>
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
 pub enum EnableExtension {
     F16,
@@ -444,7 +447,10 @@ impl LanguageExtensionName {
     }
 }
 
-/// Language extensions that can be `require`d <https://www.w3.org/TR/WGSL/#syntax-enable_extension_name>
+#[expect(clippy::doc_paragraphs_missing_punctuation, reason = "false positive")]
+/// Language extensions that can be `require`d.
+///
+/// Source: <https://www.w3.org/TR/WGSL/#syntax-enable_extension_name>
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
 pub enum LanguageExtension {
     ReadonlyAndReadwriteStorageTextures,
@@ -472,7 +478,7 @@ ast_node! {
     relative: Option<ImportRelative>;
 }
 impl Path {
-    /// Returns a list of identifiers
+    /// Returns a list of identifiers.
     pub fn segments(&self) -> impl Iterator<Item = SyntaxToken> {
         self.syntax()
             .children_with_tokens()